--- kernel/src/memory/frame_allocator/mod.rs
@@ -1,6 +1,0 @@
-#![allow(unused)]
-mod bitmap;
-mod bump;
-
-pub use bitmap::BitmapFrameAllocator;
-pub use bump::BumpFrameAllocator;
