+++ kernel/src/memory/mod.rs
@@ -1,15 +1,15 @@
-mod frame_allocator;
 mod heap_allocator;
 mod map;
 mod paging;
 
 use core::sync::atomic::{AtomicBool, Ordering};
 
+use kernel_shared::memory::frame_alloc::bitmap::BitmapFrameAllocator;
 use multiboot::{ElfSymbols, Module};
 pub use paging::*;
-use x86_64::structures::Frame;
+use x86_64::{align_up_to_page, structures::Frame};
 
-use crate::{memory::frame_allocator::BitmapFrameAllocator, BootInfo};
+use crate::BootInfo;
 
 macro_rules! log_mapping {
     ($string:expr, start: $start:expr, end: $end:expr $(, $args:expr)*) => {
@@ -18,8 +18,8 @@ macro_rules! log_mapping {
             $($args,)*
             $start,
             $end - $start,
-            crate::memory::align_down($start as usize, crate::memory::PAGE_SIZE),
-            crate::memory::align_up($end as usize, crate::memory::PAGE_SIZE),
+            x86_64::align_down_to_page($start as usize),
+            x86_64::align_up_to_page($end as usize),
         );
     };
     ($string:expr, start: $start:expr, len: $len:expr $(, $args:expr)*) => {
@@ -31,32 +31,6 @@ use log_mapping;
 
 /// Size of a page in bytes
 pub const PAGE_SIZE: usize = 4096;
-
-pub trait FrameAllocator {
-    /// Finds a free frame to allocate and return
-    fn allocate_frame(&mut self) -> Option<Frame>;
-
-    /// Frees a given frame
-    fn deallocate_frame(&mut self, frame: Frame);
-}
-
-/// Align downwards - returns the greatest _x_ with alignment `align`
-/// such that _x_ <= addr. `align` must be power of 2
-pub fn align_down(addr: usize, align: usize) -> usize {
-    if align.is_power_of_two() {
-        addr & !(align - 1)
-    } else if align == 0 {
-        addr
-    } else {
-        panic!("`align` must be power of two")
-    }
-}
-
-/// Align upwards - returns the smallest _x_ with alignment `align`
-/// such that _x_ >= addr. `align` must be power of 2
-pub fn align_up(addr: usize, align: usize) -> usize {
-    align_down(addr + align - 1, align)
-}
 
 static INIT_CALLED: AtomicBool = AtomicBool::new(false);
 
@@ -86,11 +60,10 @@ pub fn init(bootinfo: &BootInfo, initrd: &Module) {
 
     // find end address of important data structures, and map frame allocator after them
     // TODO: smarter placement of allocator backing bitmaps
-    let alloc_start_addr = align_up(
+    let alloc_start_addr = align_up_to_page(
         multiboot_end
             .max(kernel_end as usize)
             .max(initrd.end as usize),
-        PAGE_SIZE,
     );
     let (mut frame_allocator, (frame_start, frame_end)) =
         BitmapFrameAllocator::new(alloc_start_addr, bootinfo.memory_map.unwrap().entries);
