+++ kernel_shared/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 crabstd = { path = "../crabstd", default-features = false }
 x86_64 = { path = "../x86_64" }
+multiboot = { path = "../multiboot" }
 bitflags = "2.5.0"
 log = "0.4.21"
