+++ kernel_shared/src/lib.rs
@@ -1,5 +1,6 @@
 #![no_std]
 
 pub mod logger;
+pub mod memory;
 pub mod port;
 pub mod serial;
