+++ kernel_shared/src/memory/mod.rs
@@ -1,0 +1,1 @@
+pub mod frame_alloc;
